+++ src/smt/theory_str_noodler/formula.h
@@ -271,7 +271,7 @@ namespace smt::noodler {
     /**
      * Apply the given substitution. Variables are assumed to be named uniquely in the formula.
      */
-    LenNode substitute_free_vars_for_concrete_values(const LenNode& formula, const std::map<BasicTerm, int64_t> values);
+    LenNode substitute_free_vars_for_concrete_values(const LenNode& formula, const std::map<BasicTerm, int> values);
 
     //----------------------------------------------------------------------------------------------------------------------------------
